requirements.txt
    -i https://pypi.org/simple
    certifi==2019.9.11
    chardet==3.0.4
    click==7.0
    flask==1.0.2
    future==0.18.1
    idna==2.8
    itsdangerous==1.1.0
    jinja2==2.10.3
    line-bot-sdk==1.14.0
    markupsafe==1.1.1
    pygraphviz==1.5
    python-dotenv==0.10.3
    requests==2.22.0
    six==1.12.0
    transitions==0.6.9
    urllib3==1.25.6
    werkzeug==0.16.0
    openai==0.25.0
    opencc-python-reimplemented==0.1.6
    google-cloud-vision==3.1.4